consolidate constant files
