+++ src/common/constants.js
@@ -1,1 +1,3 @@
 export const COLORS = ['#386DB0', '#F7A317', '#319106', '#E56E60', '#F35E00'];
+
+export const PAGE_SIZES = [10, 20, 50];
