--- src/constants/table.js
@@ -1,1 +1,0 @@
-export const PAGE_SIZES = [10, 20, 50];
